obsolete/remove types
